# Calculate and print the average time