--- a/paddleseg/core/train.py
+++ b/paddleseg/core/train.py
@@ -20,7 +20,7 @@ import shutil
 import paddle
 import paddle.nn.functional as F
 
-from paddleseg.utils import Timer, calculate_eta, resume, logger
+from paddleseg.utils import TimeAverager, calculate_eta, resume, logger
 from paddleseg.core.val import evaluate
 
 
@@ -112,16 +112,15 @@ def train(model,
         from visualdl import LogWriter
         log_writer = LogWriter(save_dir)
 
-    timer = Timer()
     avg_loss = 0.0
     avg_loss_list = []
     iters_per_epoch = len(batch_sampler)
     best_mean_iou = -1.0
     best_model_iter = -1
-    train_reader_cost = 0.0
-    train_batch_cost = 0.0
+    reader_cost_averager = TimeAverager()
+    batch_cost_averager = TimeAverager()
     save_models = deque()
-    timer.start()
+    batch_start = time.time()
 
     iter = start_iter
     while iter < iters:
@@ -129,7 +128,7 @@ def train(model,
             iter += 1
             if iter > iters:
                 break
-            train_reader_cost += timer.elapsed_time()
+            reader_cost_averager.record(time.time() - batch_start)
             images = data[0]
             labels = data[1].astype('int64')
             edges = None
@@ -160,24 +159,24 @@ def train(model,
             else:
                 for i in range(len(loss_list)):
                     avg_loss_list[i] += loss_list[i]
-            train_batch_cost += timer.elapsed_time()
+            batch_cost_averager.record(
+                    time.time() - batch_start,
+                    num_samples=batch_size)
 
             if (iter) % log_iters == 0 and local_rank == 0:
                 avg_loss /= log_iters
                 avg_loss_list = [
                     l.numpy()[0] / log_iters for l in avg_loss_list
                 ]
-                avg_train_reader_cost = train_reader_cost / log_iters
-                avg_train_batch_cost = train_batch_cost / log_iters
-                train_reader_cost = 0.0
-                train_batch_cost = 0.0
                 remain_iters = iters - iter
+                avg_train_batch_cost = batch_cost_averager.get_average()
+                avg_train_reader_cost = reader_cost_averager.get_average()
                 eta = calculate_eta(remain_iters, avg_train_batch_cost)
                 logger.info(
-                    "[TRAIN] epoch={}, iter={}/{}, loss={:.4f}, lr={:.6f}, batch_cost={:.4f}, reader_cost={:.4f} | ETA {}"
+                    "[TRAIN] epoch={}, iter={}/{}, loss={:.4f}, lr={:.6f}, batch_cost={:.4f}, reader_cost={:.4f}, ips={:.4f} samples/sec | ETA {}"
                     .format((iter - 1) // iters_per_epoch + 1, iter, iters,
                             avg_loss, lr, avg_train_batch_cost,
-                            avg_train_reader_cost, eta))
+                            avg_train_reader_cost, batch_cost_averager.get_ips_average(), eta))
                 if use_vdl:
                     log_writer.add_scalar('Train/loss', avg_loss, iter)
                     # Record all losses if there are more than 2 losses.
@@ -196,6 +195,8 @@ def train(model,
                                           avg_train_reader_cost, iter)
                 avg_loss = 0.0
                 avg_loss_list = []
+                reader_cost_averager.reset()
+                batch_cost_averager.reset()
 
             if (iter % save_interval == 0
                     or iter == iters) and (val_dataset is not None):
@@ -233,7 +234,7 @@ def train(model,
                     if use_vdl:
                         log_writer.add_scalar('Evaluate/mIoU', mean_iou, iter)
                         log_writer.add_scalar('Evaluate/Acc', acc, iter)
-            timer.restart()
+            batch_start = time.time()
 
     # Calculate flops.
     if local_rank == 0: