--- a/paddleseg/utils/__init__.py
+++ b/paddleseg/utils/__init__.py
@@ -17,5 +17,5 @@ from . import download
 from . import metrics
 from .env import seg_env, get_sys_env
 from .utils import *
-from .timer import Timer, calculate_eta
+from .timer import TimeAverager, calculate_eta
 from . import visualize