根据benchmark标准化，增加ips统计
## train
- before
```
2021-01-19 07:02:41 [INFO]	[TRAIN] epoch=1, iter=5/100, loss=2.8735, lr=0.009639, batch_cost=0.8913, reader_cost=0.2592 | ETA 00:01:24
2021-01-19 07:02:44 [INFO]	[TRAIN] epoch=1, iter=10/100, loss=2.2584, lr=0.009186, batch_cost=0.6143, reader_cost=0.0000 | ETA 00:00:55
2021-01-19 07:02:47 [INFO]	[TRAIN] epoch=1, iter=15/100, loss=1.6819, lr=0.008731, batch_cost=0.6180, reader_cost=0.0000 | ETA 00:00:52
2021-01-19 07:02:50 [INFO]	[TRAIN] epoch=1, iter=20/100, loss=2.2924, lr=0.008272, batch_cost=0.5972, reader_cost=0.0000 | ETA 00:00:47
2021-01-19 07:02:53 [INFO]	[TRAIN] epoch=1, iter=25/100, loss=1.4582, lr=0.007811, batch_cost=0.5986, reader_cost=0.0000 | ETA 00:00:44
2021-01-19 07:02:56 [INFO]	[TRAIN] epoch=1, iter=30/100, loss=1.5533, lr=0.007347, batch_cost=0.5996, reader_cost=0.0000 | ETA 00:00:41
2021-01-19 07:02:59 [INFO]	[TRAIN] epoch=1, iter=35/100, loss=1.4564, lr=0.006880, batch_cost=0.5977, reader_cost=0.0000 | ETA 00:00:38
2021-01-19 07:03:02 [INFO]	[TRAIN] epoch=1, iter=40/100, loss=1.5977, lr=0.006409, batch_cost=0.5996, reader_cost=0.0000 | ETA 00:00:35
2021-01-19 07:03:05 [INFO]	[TRAIN] epoch=1, iter=45/100, loss=2.1524, lr=0.005934, batch_cost=0.5987, reader_cost=0.0000 | ETA 00:00:32
2021-01-19 07:03:08 [INFO]	[TRAIN] epoch=1, iter=50/100, loss=1.9311, lr=0.005455, batch_cost=0.5986, reader_cost=0.0001 | ETA 00:00:29
2021-01-19 07:03:11 [INFO]	[TRAIN] epoch=1, iter=55/100, loss=1.8286, lr=0.004971, batch_cost=0.6020, reader_cost=0.0000 | ETA 00:00:27
2021-01-19 07:03:14 [INFO]	[TRAIN] epoch=1, iter=60/100, loss=1.9385, lr=0.004482, batch_cost=0.6111, reader_cost=0.0000 | ETA 00:00:24
```
- after
```
2021-01-19 06:22:40 [INFO]	[TRAIN] epoch=1, iter=5/200, loss=3.2654, lr=0.009820, batch_cost=0.8908, reader_cost=0.23125, ips=2.2453 samples/sec | ETA 00:02:53
2021-01-19 06:22:43 [INFO]	[TRAIN] epoch=1, iter=10/200, loss=2.5065, lr=0.009594, batch_cost=0.6130, reader_cost=0.00006, ips=3.2627 samples/sec | ETA 00:01:56
2021-01-19 06:22:46 [INFO]	[TRAIN] epoch=1, iter=15/200, loss=1.8656, lr=0.009368, batch_cost=0.6032, reader_cost=0.00006, ips=3.3154 samples/sec | ETA 00:01:51
2021-01-19 06:22:49 [INFO]	[TRAIN] epoch=1, iter=20/200, loss=1.9855, lr=0.009141, batch_cost=0.5995, reader_cost=0.00007, ips=3.3361 samples/sec | ETA 00:01:47
2021-01-19 06:22:52 [INFO]	[TRAIN] epoch=1, iter=25/200, loss=1.6710, lr=0.008913, batch_cost=0.5994, reader_cost=0.00006, ips=3.3366 samples/sec | ETA 00:01:44
2021-01-19 06:22:55 [INFO]	[TRAIN] epoch=1, iter=30/200, loss=1.7677, lr=0.008685, batch_cost=0.5915, reader_cost=0.00005, ips=3.3814 samples/sec | ETA 00:01:40
2021-01-19 06:22:58 [INFO]	[TRAIN] epoch=1, iter=35/200, loss=1.6532, lr=0.008456, batch_cost=0.5997, reader_cost=0.00005, ips=3.3350 samples/sec | ETA 00:01:38
2021-01-19 06:23:01 [INFO]	[TRAIN] epoch=1, iter=40/200, loss=2.0951, lr=0.008227, batch_cost=0.6024, reader_cost=0.00006, ips=3.3202 samples/sec | ETA 00:01:36
2021-01-19 06:23:04 [INFO]	[TRAIN] epoch=1, iter=45/200, loss=1.6151, lr=0.007996, batch_cost=0.5931, reader_cost=0.00004, ips=3.3721 samples/sec | ETA 00:01:31
2021-01-19 06:23:07 [INFO]	[TRAIN] epoch=1, iter=50/200, loss=1.7092, lr=0.007765, batch_cost=0.6057, reader_cost=0.00026, ips=3.3021 samples/sec | ETA 00:01:30
2021-01-19 06:23:10 [INFO]	[TRAIN] epoch=1, iter=55/200, loss=1.7479, lr=0.007533, batch_cost=0.5985, reader_cost=0.00005, ips=3.3417 samples/sec | ETA 00:01:26
2021-01-19 06:23:13 [INFO]	[TRAIN] epoch=1, iter=60/200, loss=1.8309, lr=0.007301, batch_cost=0.5953, reader_cost=0.00005, ips=3.3597 samples/sec | ETA 00:01:23
```
## eval
- before
```
500/500 [==============================] - 118s 236ms/step - batch_cost: 0.2354 - reader cost: 0.0198
```
- after
```
500/500 [==============================] - 119s 237ms/step - batch_cost: 0.2371 - reader cost: 0.0421
```